Disable std::max and std::min overloading

+++ aten/src/ATen/native/quantized/fake_quant_per_tensor_affine.cpp
@@ -106,9 +106,10 @@ int64_t _get_zero_point_from_tensor(
     int64_t quant_min,
     int64_t quant_max,
     bool is_forward) {
-  float zero_point_fp = zero_point[0].item<float>();
+  double zero_point_fp = zero_point[0].item<double>();
   zero_point_fp = is_forward ? std::nearbyint(zero_point_fp) : zero_point_fp + 0.5f;
-  float zero_point_clamped = std::min(std::max(zero_point_fp, quant_min), quant_max);
+  double zero_point_clamped = std::min(std::max(zero_point_fp, static_cast<double>(quant_min)),
+                                       static_cast<double>(quant_max));
   return static_cast<int64_t>(zero_point_clamped);
 }
